Update docs for swappable models
