+++ docs/advanced/custom_formatters.rst
@@ -15,7 +15,6 @@ The default formatter function is set to :func:`salesman.core.utils.format_price
 a value formated with two decimal places.
 
 .. tip::
-
     Price formatter can be used *(among other things)* to display a price with a symbol or even a
     converted price to another currency based on the ``request`` that's received in the ``context``.
 
@@ -28,7 +27,6 @@ that points to your custom formatter function.
 Your custom function should accept a ``value`` argument of type ``Decimal`` and a ``context``
 dictionary that contains additional render data like ``request`` and either the ``basket`` or
 ``order`` object. The function should return a formatted price as a string.
-
 
 
 .. _admin-json-formatter:
@@ -47,20 +45,3 @@ to a function in ``SALESMAN_ADMIN_JSON_FORMATTER`` setting.
 Your custom formatter should accept a dictionary ``value`` and return the HTML string.
 It also receives a ``context`` dictionary with additional context. Either an ``order`` or
 ``order_item`` boolean will be passed in depending on the formatting location.
-
-
-.. _admin-customer-formatter:
-
-Admin Customer formatter
-========================
-
-To display customer on an order, a formatter function is used. The default function
-:func:`salesman.admin.utils.format_customer` returns the link to the default User models
-for both the Django and Wagtail admins. You can override the Customer formatter by providing
-a dotted path to a function in ``SALESMAN_ADMIN_CUSTOMER_FORMATTER`` setting.
-
-.. literalinclude:: /../salesman/admin/utils.py
-    :pyobject: format_customer
-
-Your custom formatter should accept a ``user`` instnance and return a string. It also receives a
-``context`` dictionary with additional context.
