+++ example/shop/payment/__init__.py
@@ -1,0 +1,3 @@
+from .advance import *  # noqa
+from .credit_card import *  # noqa
+from .on_delivery import *  # noqa
